BenchmarkResult: update type annotation for data/times
with the SQLAlchemy Numeric option this is documented
to return decimal.Decimal, not float. That is where
a lot of confusion came from. Woof.

+++ conbench/entities/benchmark_result.py
@@ -70,10 +70,10 @@ class BenchmarkResult(Base, EntityMixin):
 
     # These can be empty lists. An item in the list is of type float, which
     # includes `math.nan` as valid value.
-    data: Mapped[Optional[List[float]]] = Nullable(
+    data: Mapped[Optional[List[Decimal]]] = Nullable(
         postgresql.ARRAY(s.Numeric), default=[]
     )
-    times: Mapped[Optional[List[float]]] = Nullable(
+    times: Mapped[Optional[List[Decimal]]] = Nullable(
         postgresql.ARRAY(s.Numeric), default=[]
     )
